+++ tests/integration/models/reddit/test_subreddit.py
@@ -712,27 +712,14 @@ class TestSubredditFlairTemplates(IntegrationTest):
         for flair_template in templates:
             assert flair_template["id"]
 
-    @mock.patch("time.sleep", return_value=None)
-    def test_add(self, _):
+    def test_add(self):
         self.reddit.read_only = False
         with self.recorder.use_cassette(
             "TestSubredditFlairTemplates.test_add"
         ):
-            for i in range(101):
-                self.subreddit.flair.templates.add(
-                    "PRAW{}".format(i), css_class="myCSS"
-                )
-
-    @mock.patch("time.sleep", return_value=None)
-    def test_add_v2(self, _):
-        self.reddit.read_only = False
-        with self.recorder.use_cassette(
-            "TestSubredditFlairTemplates.test_add_v2"
-        ):
-            for i in range(101):
-                self.subreddit.flair.templates.add(
-                    "PRAW{}".format(i), background_color="#ABCDEF"
-                )
+            self.subreddit.flair.templates.add(
+                "PRAW", css_class="myCSS", background_color="#ABCDEF"
+            )
 
     def test_clear(self):
         self.reddit.read_only = False
@@ -784,27 +771,14 @@ class TestSubredditLinkFlairTemplates(IntegrationTest):
             assert isinstance(template["richtext"], list)
             assert all(isinstance(item, dict) for item in template["richtext"])
 
-    @mock.patch("time.sleep", return_value=None)
-    def test_add(self, _):
+    def test_add(self):
         self.reddit.read_only = False
         with self.recorder.use_cassette(
             "TestSubredditLinkFlairTemplates.test_add"
         ):
-            for i in range(101):
-                self.subreddit.flair.link_templates.add(
-                    "PRAW{}".format(i), css_class="myCSS"
-                )
-
-    @mock.patch("time.sleep", return_value=None)
-    def test_add_v2(self, _):
-        self.reddit.read_only = False
-        with self.recorder.use_cassette(
-            "TestSubredditLinkFlairTemplates.test_add_v2"
-        ):
-            for i in range(101):
-                self.subreddit.flair.link_templates.add(
-                    "PRAW{}".format(i), text_color="light"
-                )
+            self.subreddit.flair.link_templates.add(
+                "PRAW", css_class="myCSS", text_color="light"
+            )
 
     def test_clear(self):
         self.reddit.read_only = False